+++ wallet/api/client_check_transaction_test.go
@@ -4,7 +4,9 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"sync"
 	"testing"
+	"time"
 
 	"code.vegaprotocol.io/vega/libs/jsonrpc"
 	vgrand "code.vegaprotocol.io/vega/libs/rand"
@@ -23,6 +25,7 @@ func TestClientCheckTransaction(t *testing.T) {
 	t.Run("Documentation matches the code", testClientCheckTransactionSchemaCorrect)
 	t.Run("Checking a transaction with invalid params fails", testCheckingTransactionWithInvalidParamsFails)
 	t.Run("Checking a transaction with valid params succeeds", testCheckingTransactionWithValidParamsSucceeds)
+	t.Run("Checking a transaction in parallel blocks on same party but not on different parties", testCheckingTransactionInParallelBlocksOnSamePartyButNotOnDifferentParties)
 	t.Run("Checking a transaction without the needed permissions check the transaction", testCheckingTransactionWithoutNeededPermissionsDoesNotCheckTransaction)
 	t.Run("Refusing the checking of a transaction does not check the transaction", testRefusingCheckingOfTransactionDoesNotCheckTransaction)
 	t.Run("Cancelling the review does not check the transaction", testCancellingTheReviewDoesNotCheckTransaction)
@@ -171,6 +174,204 @@ func testCheckingTransactionWithValidParamsSucceeds(t *testing.T) {
 	assert.Nil(t, errorDetails)
 	require.NotEmpty(t, result)
 	assert.NotEmpty(t, result.Transaction)
+}
+
+func testCheckingTransactionInParallelBlocksOnSamePartyButNotOnDifferentParties(t *testing.T) {
+	// setup
+
+	// Use channels to orchestrate requests.
+	sendSecondRequests := make(chan interface{})
+	sendThirdRequests := make(chan interface{})
+	waitForSecondRequestToExit := make(chan interface{})
+	waitForThirdRequestToExit := make(chan interface{})
+
+	hostname := vgrand.RandomStr(5)
+
+	// One context for each request.
+	r1Ctx, r1TraceID := clientContextForTest()
+	r2Ctx, _ := clientContextForTest()
+	r3Ctx, r3TraceID := clientContextForTest()
+
+	// A wallet with 2 keys to have 2 different parties.
+	wallet1 := walletWithPerms(t, hostname, wallet.Permissions{
+		PublicKeys: wallet.PublicKeysPermission{
+			Access:      wallet.ReadAccess,
+			AllowedKeys: nil,
+		},
+	})
+	kp1, err := wallet1.GenerateKeyPair(nil)
+	require.NoError(t, err)
+	kp2, err := wallet1.GenerateKeyPair(nil)
+	require.NoError(t, err)
+
+	// We can have a single connection as the implementation only cares about the
+	// party.
+	connectedWallet, err := api.NewConnectedWallet(hostname, wallet1)
+	require.NoError(t, err)
+
+	// Some mock data. Their value is irrelevant to test parallelism, so we recycle
+	// them.
+	spamStats := types.SpamStatistics{
+		ChainID:           vgrand.RandomStr(5),
+		LastBlockHeight:   100,
+		Proposals:         &types.SpamStatistic{MaxForEpoch: 1},
+		NodeAnnouncements: &types.SpamStatistic{MaxForEpoch: 1},
+		Delegations:       &types.SpamStatistic{MaxForEpoch: 1},
+		Transfers:         &types.SpamStatistic{MaxForEpoch: 1},
+		Votes:             &types.VoteSpamStatistics{MaxForEpoch: 1},
+		PoW: &types.PoWStatistics{
+			PowBlockStates: []types.PoWBlockState{{}},
+		},
+	}
+	pow := &commandspb.ProofOfWork{
+		Tid:   vgrand.RandomStr(5),
+		Nonce: 12345678,
+	}
+
+	// Setting up the mocked calls. The second request shouldn't trigger any of
+	// them, since it should be rejected because it uses the same party as the
+	// first request, which only unblock at the end.
+	handler := newCheckTransactionHandler(t)
+
+	gomock.InOrder(
+		// First request.
+		handler.spam.EXPECT().GenerateProofOfWork(kp1.PublicKey(), &spamStats).Times(1).Return(pow, nil),
+		// Third request.
+		handler.spam.EXPECT().GenerateProofOfWork(kp2.PublicKey(), &spamStats).Times(1).Return(pow, nil),
+	)
+	gomock.InOrder(
+		// First request.
+		handler.interactor.EXPECT().NotifyInteractionSessionBegan(r1Ctx, r1TraceID, api.TransactionReviewWorkflow, uint8(2)).Times(1).Return(nil),
+		// Third request.
+		handler.interactor.EXPECT().NotifyInteractionSessionBegan(r3Ctx, r3TraceID, api.TransactionReviewWorkflow, uint8(2)).Times(1).Return(nil),
+	)
+	gomock.InOrder(
+		// Third request is expected before because the first request get unblocked
+		// when the third request finishes.
+		handler.interactor.EXPECT().NotifyInteractionSessionEnded(r3Ctx, r3TraceID).Times(1),
+		// First request.
+		handler.interactor.EXPECT().NotifyInteractionSessionEnded(r1Ctx, r1TraceID).Times(1),
+	)
+	gomock.InOrder(
+		// First request.
+		handler.interactor.EXPECT().RequestTransactionReviewForChecking(r1Ctx, r1TraceID, uint8(1), hostname, wallet1.Name(), kp1.PublicKey(), fakeTransaction, gomock.Any()).Times(1).Return(true, nil),
+		// Third request.
+		handler.interactor.EXPECT().RequestTransactionReviewForChecking(r3Ctx, r3TraceID, uint8(1), hostname, wallet1.Name(), kp2.PublicKey(), fakeTransaction, gomock.Any()).Times(1).Return(true, nil),
+	)
+	gomock.InOrder(
+		// First request.
+		handler.nodeSelector.EXPECT().Node(r1Ctx, gomock.Any()).Times(1).Return(handler.node, nil),
+		// Third request.
+		handler.nodeSelector.EXPECT().Node(r3Ctx, gomock.Any()).Times(1).Return(handler.node, nil),
+	)
+	gomock.InOrder(
+		// First request.
+		handler.walletStore.EXPECT().GetWallet(r1Ctx, wallet1.Name()).Times(1).Return(wallet1, nil),
+		// Second request.
+		handler.walletStore.EXPECT().GetWallet(r2Ctx, wallet1.Name()).Times(1).DoAndReturn(func(_ context.Context, _ string) (wallet.Wallet, error) {
+			close(sendThirdRequests)
+			return wallet1, nil
+		}),
+		// Third request.
+		handler.walletStore.EXPECT().GetWallet(r3Ctx, wallet1.Name()).Times(1).Return(wallet1, nil),
+	)
+	gomock.InOrder(
+		// First request.
+		handler.node.EXPECT().SpamStatistics(r1Ctx, kp1.PublicKey()).Times(1).Return(spamStats, nil),
+		// Third request.
+		handler.node.EXPECT().SpamStatistics(r3Ctx, kp2.PublicKey()).Times(1).Return(spamStats, nil),
+	)
+	gomock.InOrder(
+		// First request.
+		handler.spam.EXPECT().CheckSubmission(gomock.Any(), &spamStats).Times(1).Return(nil),
+		// Third request.
+		handler.spam.EXPECT().CheckSubmission(gomock.Any(), &spamStats).Times(1).Return(nil),
+	)
+	gomock.InOrder(
+		// First request.
+		handler.interactor.EXPECT().NotifySuccessfulRequest(r1Ctx, r1TraceID, uint8(2), api.TransactionSuccessfullyChecked).Times(1).Do(func(_ context.Context, _ string, _ uint8, _ string) {
+			// Unblock the second and third requests, and trigger the signing.
+			close(sendSecondRequests)
+			<-waitForSecondRequestToExit
+		}),
+		// Third request.
+		handler.interactor.EXPECT().NotifySuccessfulRequest(r3Ctx, r3TraceID, uint8(2), api.TransactionSuccessfullyChecked).Times(1),
+	)
+	gomock.InOrder(
+		// First request.
+		handler.node.EXPECT().CheckTransaction(r1Ctx, gomock.Any()).AnyTimes().Return(nil),
+		// Third request.
+		handler.node.EXPECT().CheckTransaction(r3Ctx, gomock.Any()).AnyTimes().Return(nil),
+	)
+	gomock.InOrder(
+		// First request.
+		handler.interactor.EXPECT().Log(r1Ctx, r1TraceID, gomock.Any(), gomock.Any()).AnyTimes(),
+		// Third request.
+		handler.interactor.EXPECT().Log(r3Ctx, r3TraceID, gomock.Any(), gomock.Any()).AnyTimes(),
+	)
+
+	wg := sync.WaitGroup{}
+	wg.Add(3)
+
+	go func() {
+		defer wg.Done()
+		// when
+		result, errorDetails := handler.handle(t, r1Ctx, api.ClientCheckTransactionParams{
+			PublicKey:   kp1.PublicKey(),
+			Transaction: testTransaction(t),
+		}, connectedWallet)
+
+		<-waitForSecondRequestToExit
+		<-waitForThirdRequestToExit
+
+		// then
+		assert.Nil(t, errorDetails)
+		require.NotEmpty(t, result)
+		assert.NotEmpty(t, result.Transaction)
+	}()
+
+	go func() {
+		defer wg.Done()
+
+		// Closing this resume, unblock the first request.
+		defer close(waitForSecondRequestToExit)
+
+		// Ensure the first request acquire the "lock" on the public key.
+		<-sendSecondRequests
+
+		// when
+		result, errorDetails := handler.handle(t, r2Ctx, api.ClientCheckTransactionParams{
+			PublicKey:   kp1.PublicKey(),
+			Transaction: testTransaction(t),
+		}, connectedWallet)
+
+		// then
+		assert.NotNil(t, errorDetails)
+		assertRequestNotPermittedError(t, errorDetails, fmt.Errorf("this public key %q is already in use, retry later", kp1.PublicKey()))
+		require.Empty(t, result)
+	}()
+
+	go func() {
+		defer wg.Done()
+		defer close(waitForThirdRequestToExit)
+
+		// Ensure the first request acquire the "lock" on the public key, and
+		// we second request calls `GetWallet()` before the third request.
+		<-sendThirdRequests
+
+		// then
+		result, errorDetails := handler.handle(t, r3Ctx, api.ClientCheckTransactionParams{
+			PublicKey:   kp2.PublicKey(),
+			Transaction: testTransaction(t),
+		}, connectedWallet)
+
+		// then
+		assert.Nil(t, errorDetails)
+		require.NotEmpty(t, result)
+		assert.NotEmpty(t, result.Transaction)
+	}()
+
+	wg.Wait()
 }
 
 func testCheckingTransactionWithoutNeededPermissionsDoesNotCheckTransaction(t *testing.T) {
@@ -596,8 +797,13 @@ func newCheckTransactionHandler(t *testing.T) *checkTransactionHandler {
 	walletStore := mocks.NewMockWalletStore(ctrl)
 	node := nodemocks.NewMockNode(ctrl)
 
+	requestController := api.NewRequestController(
+		api.WithMaximumAttempt(1),
+		api.WithIntervalDelayBetweenRetries(1*time.Second),
+	)
+
 	return &checkTransactionHandler{
-		ClientCheckTransaction: api.NewClientCheckTransaction(walletStore, interactor, nodeSelector, proofOfWork),
+		ClientCheckTransaction: api.NewClientCheckTransaction(walletStore, interactor, nodeSelector, proofOfWork, requestController),
 		ctrl:                   ctrl,
 		nodeSelector:           nodeSelector,
 		interactor:             interactor,